cfg, ns: factor out qrtr_set_address function
This function in qrtr-ns was largely copied from the main
function of qrtr-cfg. We should just factor it out so there is
one implementation instead.
